e2e test improvements
